requirements.txt
pandas
nicegui
plotly
fpdf2
uvicorn
pyinstaller
pywebview
